requirements.txt
gitpython
python-dateutil